+++ app/client/src/workers/Linting/constants.ts
@@ -1,0 +1,29 @@
+import { ECMA_VERSION } from "@shared/ast";
+import { LintOptions } from "jshint";
+
+export const lintOptions = (globalData: Record<string, boolean>) =>
+  ({
+    indent: 2,
+    esversion: ECMA_VERSION,
+    eqeqeq: false, // Not necessary to use ===
+    curly: false, // Blocks can be added without {}, eg if (x) return true
+    freeze: true, // Overriding inbuilt classes like Array is not allowed
+    undef: true, // Undefined variables should be reported as error
+    forin: false, // Doesn't require filtering for..in loops with obj.hasOwnProperty()
+    noempty: false, // Empty blocks are allowed
+    strict: false, // We won't force strict mode
+    unused: "strict", // Unused variables are not allowed
+    asi: true, // Tolerate Automatic Semicolon Insertion (no semicolons)
+    boss: true, // Tolerate assignments where comparisons would be expected
+    evil: false, // Use of eval not allowed
+    funcscope: true, // Tolerate variable definition inside control statements
+    sub: true, // Don't force dot notation
+    expr: true, // suppresses warnings about the use of expressions where normally you would expect to see assignments or function calls
+    // environments
+    browser: true,
+    worker: true,
+    mocha: false,
+    // global values
+    globals: globalData,
+    loopfunc: true,
+  } as LintOptions);
