+++ app/client/src/workers/Linting/index.ts
@@ -1,0 +1,121 @@
+import {
+  getEntityNameAndPropertyPath,
+  isATriggerPath,
+  isJSAction,
+} from "ce/workers/Evaluation/evaluationUtils";
+import { DataTree } from "entities/DataTree/dataTreeFactory";
+import { get, set } from "lodash";
+import { LintErrors } from "reducers/lintingReducers/lintErrorsReducers";
+import { createEvaluationContext } from "workers/Evaluation/evaluate";
+import { getActionTriggerFunctionNames } from "workers/Evaluation/fns";
+import { lintBindingPath, lintTriggerPath, pathRequiresLinting } from "./utils";
+
+export function getlintErrorsFromTree(
+  pathsToLint: string[],
+  unEvalTree: DataTree,
+  cloudHosting: boolean,
+): LintErrors {
+  const lintTreeErrors: LintErrors = {};
+
+  const evalContext = createEvaluationContext({
+    dataTree: unEvalTree,
+    resolvedFunctions: {},
+    isTriggerBased: false,
+    skipEntityFunctions: true,
+  });
+
+  const platformFnNamesMap = Object.values(
+    getActionTriggerFunctionNames(cloudHosting),
+  ).reduce(
+    (acc, name) => ({ ...acc, [name]: true }),
+    {} as { [x: string]: boolean },
+  );
+  Object.assign(evalContext, platformFnNamesMap);
+
+  const evalContextWithoutFunctions = createEvaluationContext({
+    dataTree: unEvalTree,
+    resolvedFunctions: {},
+    isTriggerBased: true,
+    skipEntityFunctions: true,
+  });
+
+  // trigger paths
+  const triggerPaths = new Set<string>();
+  // Certain paths, like JS Object's body are binding paths where appsmith functions are needed in the global data
+  const bindingPathsRequiringFunctions = new Set<string>();
+
+  pathsToLint.forEach((fullPropertyPath) => {
+    const { entityName, propertyPath } = getEntityNameAndPropertyPath(
+      fullPropertyPath,
+    );
+    const entity = unEvalTree[entityName];
+    const unEvalPropertyValue = (get(
+      unEvalTree,
+      fullPropertyPath,
+    ) as unknown) as string;
+    // remove all lint errors from path
+    set(lintTreeErrors, `["${fullPropertyPath}"]`, []);
+
+    // We are only interested in paths that require linting
+    if (!pathRequiresLinting(unEvalTree, entity, fullPropertyPath)) return;
+    if (isATriggerPath(entity, propertyPath))
+      return triggerPaths.add(fullPropertyPath);
+    if (isJSAction(entity))
+      return bindingPathsRequiringFunctions.add(`${entityName}.body`);
+    const lintErrors = lintBindingPath(
+      unEvalPropertyValue,
+      entity,
+      fullPropertyPath,
+      evalContextWithoutFunctions,
+    );
+    set(lintTreeErrors, `["${fullPropertyPath}"]`, lintErrors);
+  });
+
+  if (triggerPaths.size || bindingPathsRequiringFunctions.size) {
+    // we only create GLOBAL_DATA_WITH_FUNCTIONS if there are paths requiring it
+    // In trigger based fields, functions such as showAlert, storeValue, etc need to be added to the global data
+
+    // lint binding paths that need GLOBAL_DATA_WITH_FUNCTIONS
+    if (bindingPathsRequiringFunctions.size) {
+      bindingPathsRequiringFunctions.forEach((fullPropertyPath) => {
+        const { entityName } = getEntityNameAndPropertyPath(fullPropertyPath);
+        const entity = unEvalTree[entityName];
+        const unEvalPropertyValue = (get(
+          unEvalTree,
+          fullPropertyPath,
+        ) as unknown) as string;
+        // remove all lint errors from path
+        set(lintTreeErrors, `["${fullPropertyPath}"]`, []);
+        const lintErrors = lintBindingPath(
+          unEvalPropertyValue,
+          entity,
+          fullPropertyPath,
+          evalContext,
+        );
+        set(lintTreeErrors, `["${fullPropertyPath}"]`, lintErrors);
+      });
+    }
+
+    // Lint triggerPaths
+    if (triggerPaths.size) {
+      triggerPaths.forEach((triggerPath) => {
+        const { entityName } = getEntityNameAndPropertyPath(triggerPath);
+        const entity = unEvalTree[entityName];
+        const unEvalPropertyValue = (get(
+          unEvalTree,
+          triggerPath,
+        ) as unknown) as string;
+        // remove all lint errors from path
+        set(lintTreeErrors, `["${triggerPath}"]`, []);
+        const lintErrors = lintTriggerPath(
+          unEvalPropertyValue,
+          entity,
+          evalContext,
+        );
+        set(lintTreeErrors, `["${triggerPath}"]`, lintErrors);
+      });
+    }
+  }
+
+  return lintTreeErrors;
+}
